requirements.txt
requests~=2.25.1
pytest~=6.2.0
PyYAML~=5.4.1
xlrd~=2.0.1